feat: Implement Battle Pass system with login, kill tracking, and reward claiming
- Added battle_pass.lua to handle player login, kills, and extended opcodes for Battle Pass interactions.
- Introduced client_config.lua to send client configuration on login.
- Created donate_bless.lua to reapply blessings for premium accounts on login based on donation tier.
- Implemented donate_bonus_kill.lua to provide experience and bonus loot based on donation tier upon monster kills.
- Developed outfit_transform.lua to manage player outfit transformations and associated effects.
- Added quest_extended_opcode.lua to handle quest log requests via extended opcodes.
- Created autoloot_catalog_reload.lua to periodically reload the autoloot catalog.
- Implemented battle_pass_rotation.lua to ensure the Battle Pass season is updated monthly.
- Developed battle_pass_lib.lua to manage Battle Pass data, player progress, and rewards.
- Created donate_tier.lua to define donation tiers and retrieve account donation status.
- Added quest_lib.lua to manage player quest progress and completion.
- Implemented server_config_lib.lua to read server configuration values from the database.
- Developed task_db_loader.lua to load task definitions from the database.
- Created resgatarquest.lua to allow players to claim rewards for completed quests.

+++ src/source/game.h
@@ -457,11 +457,11 @@ class Game
 			ViolationAction_t action, std::string comment, std::string statement,
 			uint32_t statementId, bool ipBanishment);
 		bool playerMoveThing(uint32_t playerId, const Position& fromPos, uint16_t spriteId,
-			int16_t fromStackpos, const Position& toPos, uint8_t count);
+			int16_t fromStackpos, const Position& toPos, uint16_t count);
 		bool playerMoveCreature(uint32_t playerId, uint32_t movingCreatureId,
 			const Position& movingCreatureOrigPos, const Position& toPos, bool delay);
 		bool playerMoveItem(uint32_t playerId, const Position& fromPos,
-			uint16_t spriteId, int16_t fromStackpos, const Position& toPos, uint8_t count);
+			uint16_t spriteId, int16_t fromStackpos, const Position& toPos, uint16_t count);
 		bool playerMove(uint32_t playerId, Direction dir);
 		bool playerCreatePrivateChannel(uint32_t playerId, ProtocolGame* pg); //CAST
 		bool playerChannelInvite(uint32_t playerId, const std::string& name);
